implemented settings usage instead of ENV variables, moved config loading and saving into a module
